requirements.txt
# Core Trading Framework
freqtrade==2024.10
python-binance==1.0.17

# Data Processing & Analysis
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0
xgboost==2.0.3

# Technical Indicators
ta-lib==0.4.28
talib-binary==0.4.28
tulipy==1.1.1

# Optimization & Hyperparameters
optuna==3.14.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1
jupyter==1.0.0
jupyterlab==4.0.9
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Development
pylint==3.0.3
black==23.12.1
pytest==7.4.3
flake8==6.1.0
